+++ src/Channels/channel.js
@@ -51,6 +51,8 @@ export class Channel {
     fadeStarted = false;
     source;
 
+    firstLoop = true;
+
     settings = {
         channel: 0,
         name: '',
@@ -115,20 +117,26 @@ export class Channel {
               "data": data
             };
             game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
         }
     }
 
     async setSbData(data,currentlyPlaying=0) {
+
+        if (this.mixer.mixer.mixerApp != undefined && this.mixer.mixer.mixerApp.rendered) {
+            const color = (data.repeat.repeat == 'single' || data.repeat.repeat == 'all') ? 'yellow' : 'black';
+            const elmnt = document.getElementById(`sbButton-${data.channel-100}`);
+            elmnt.style.borderColor = color;
+            elmnt.style.boxShadow = (color == 'yellow') ? "0 0 10px yellow" : "";
+        }
+
         this.loaded = false;
         this.stop();
         this.settings = data;
+
         this.setVolume(data.volume);
         if (data.sourceArray == undefined) data.sourceArray = await this.getSounds(data.soundData);
         this.sourceArray = data.sourceArray;
-        if (this.sourceArray == undefined || this.sourceArray[0] == undefined) return;
-        this.currentlyPlaying = currentlyPlaying;
-        this.setSource(this.sourceArray[currentlyPlaying]);
-        this.setPlaybackRate(data.playbackRate);
         if (game.user.isGM) {
             data.soundArray = this.soundArray;
 
@@ -138,7 +146,14 @@ export class Channel {
               "data": data
             };
             game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
         }
+        if (this.sourceArray == undefined || this.sourceArray[0] == undefined) return;
+        this.currentlyPlaying = currentlyPlaying;
+        this.setSource(this.sourceArray[currentlyPlaying]);
+        this.setPlaybackRate(data.playbackRate);
+        
+        
     }
 
     async getSounds(soundData) {
@@ -177,9 +192,9 @@ export class Channel {
             for (let file of files) 
             soundArray.push(file)
         }
-    
+
         //Randomize array
-        if (soundData.randomize) return this.randomizeArray(soundArray)
+        if (this.settings.randomize) return this.randomizeArray(soundArray)
         else return soundArray;
         
     }
@@ -194,7 +209,7 @@ export class Channel {
         return array;
     }
 
-    play(currentTime = undefined) {
+    async play(currentTime = undefined) {
         if (this.loaded == false || this.playing && this.paused==false && this.channelNr < 100) return;
         if (this.audioElement == undefined) {
             if (this.sourceArray == undefined || this.sourceArray.length == 0) return;
@@ -202,7 +217,16 @@ export class Channel {
         }
         if (this.channelNr >= 100) {
             if (this.settings.playbackRate.rate == undefined) this.settings.playbackRate.rate = 1;
-            this.setPlaybackRate(this.settings.playbackRate);
+            if (game.user.isGM) this.setPlaybackRate(this.settings.playbackRate);
+
+            if (this.mixer.mixer.mixerApp != undefined && this.mixer.mixer.mixerApp.rendered) {
+                const color = (this.settings.repeat.repeat == 'single' || this.settings.repeat.repeat == 'all') ? 'green' : 'black';
+                const elmnt = document.getElementById(`sbButton-${this.channelNr-100}`);
+                elmnt.style.borderColor = color;
+                elmnt.style.boxShadow = (color == 'green') ? "0 0 10px green" : "";
+            }
+
+            this.randomizeVolume();
         }
         else
             this.setPlaybackRate();
@@ -211,23 +235,37 @@ export class Channel {
                 timing = {
                     startTime: 0,
                     stopTime: 0,
+                    skipFirstTiming: false,
                     fadeIn: 0,
-                    fadeout: 0
+                    fadeOut: 0,
+                    skipFirstFade: false
                 }
-        if (this.paused == false) this.audioElement.currentTime = timing.startTime;
-        if (timing.fadeIn > 0) {
+        if (this.paused == false && (timing.skipFirstTiming == false || this.firstLoop == false)) this.audioElement.currentTime = timing.startTime;
+        if (timing.fadeIn > 0  && (timing.skipFirstFade == false || this.firstLoop == false)) {
             if (this.fadeStarted == false) this.fade(0,this.settings.volume,timing.fadeIn)
         }
+        this.firstLoop = false;
         if (currentTime != undefined) this.audioElement.currentTime = currentTime;
         if (this.context.state != 'running') return;
-        this.audioElement.play();
 
-        this.playing = true;
-        this.paused = false;
+        let playPromise = this.audioElement.play();
 
-        if (this.mixer.mixerApp != undefined && this.mixer.mixerApp.rendered) {
-            document.getElementById(`playSound-${this.channelNr}`).innerHTML = `<i class="fas fa-stop"></i>`;
-        }
+        if (playPromise !== undefined) {
+            playPromise.then(_ => {
+                
+            })
+            .catch(error => {
+              // Auto-play was prevented
+            });
+          }
+          this.playing = true;
+          this.paused = false;
+  
+          if (this.mixer.mixerApp != undefined && this.mixer.mixerApp.rendered) {
+              document.getElementById(`playSound-${this.channelNr}`).innerHTML = `<i class="fas fa-stop"></i>`;
+          }
+
+        
     }
 
     pause() {
@@ -237,12 +275,19 @@ export class Channel {
     }
     
     stop(next = true) {
+        this.firstLoop = true;
         if (this.audioElement == undefined || (this.playing == false && this.paused == false)) return
         this.audioElement.pause();
         this.audioElement.currentTime = 0;
         this.playing = false;
         this.paused = false;
         if (next) this.next();
+        if (this.channelNr >= 100 && this.mixer.mixer.mixerApp != undefined && this.mixer.mixer.mixerApp.rendered) {
+            const color = (this.settings.repeat.repeat == 'single' || this.settings.repeat.repeat == 'all') ? 'yellow' : 'black';
+            const elmnt = document.getElementById(`sbButton-${this.channelNr-100}`);
+            elmnt.style.borderColor = color;
+            elmnt.style.boxShadow = (color == 'yellow') ? "0 0 10px yellow" : "";
+        }
         if (this.mixer.mixerApp != undefined && this.mixer.mixerApp.rendered) {
             document.getElementById(`playSound-${this.channelNr}`).innerHTML = `<i class="fas fa-play"></i>`;
         }
@@ -255,6 +300,7 @@ export class Channel {
         else this.currentlyPlaying = playNr;
         if (this.currentlyPlaying > this.sourceArray.length - 1) this.currentlyPlaying = 0;
         this.setSource(this.sourceArray[this.currentlyPlaying]);
+        
         if (game.user.isGM) {
             const payload = {
               "msgType": "next",
@@ -263,6 +309,7 @@ export class Channel {
               "playing": this.playing
             };
             game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
         }
     }
 
@@ -271,11 +318,31 @@ export class Channel {
         this.currentlyPlaying--;
         if (this.currentlyPlaying < 0) this.currentlyPlaying = this.sourceArray.length - 1;
         this.setSource(this.sourceArray[this.currentlyPlaying]);
+        if (game.user.isGM) {
+            const payload = {
+              "msgType": "previous",
+              "channel": this.channelNr,
+              "currentlyPlaying": this.currentlyPlaying,
+              "playing": this.playing
+            };
+            game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
+        }
     }
 
     restart() {
         if (this.audioElement == undefined) return;
         this.audioElement.currentTime = 0;
+        if (game.user.isGM) {
+            const payload = {
+              "msgType": "restart",
+              "channel": this.channelNr,
+              "currentlyPlaying": this.currentlyPlaying,
+              "playing": this.playing
+            };
+            game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
+        }
     }
 
     setPlaybackRate(playbackRate=this.settings.playbackRate) {
@@ -295,12 +362,13 @@ export class Channel {
               playbackRate
             };
             game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
         }
         this.audioElement.playbackRate = playbackRate.rate;
         this.audioElement.preservesPitch = playbackRate.preservePitch;
     }
 
-    setVolume(volume=this.settings.volume,save=true) {
+    setVolume(volume=this.settings.volume,save=true,solo=false) {
         if (game.user.isGM) {
             const payload = {
               "msgType": "setVolume",
@@ -309,6 +377,7 @@ export class Channel {
               save
             };
             game.socket.emit(`module.soundscape`, payload);
+            if (solo == false) Hooks.call(moduleName,payload);
         }
         if (this.settings.mute) volume = 0;
         else if (save) this.settings.volume = volume;
@@ -325,6 +394,7 @@ export class Channel {
               pan
             };
             game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
         }
         this.effects.pan.set(pan);
     }
@@ -337,6 +407,7 @@ export class Channel {
               mute
             };
             game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
         }
         let volume = mute ? 0 : this.settings.volume;
         this.effects.gain.set(volume);
@@ -366,20 +437,9 @@ export class Channel {
               solo
             };
             game.socket.emit(`module.soundscape`, payload);
+            Hooks.call(moduleName,payload);
         }
         this.settings.solo = solo;
-
-        let channel = this.master ? 'master' : this.channelNr + 1;
-
-        const msg = {
-            module: 'Soundscape',
-            channel: channel,
-            mode: 'setSolo',
-            mute: this.settings.mute,
-            solo: solo,
-            link: this.settings.link
-        }
-        Hooks.call("Soundscape",msg);
     }
 
     getSolo() {
@@ -389,17 +449,14 @@ export class Channel {
     setLink(link) {
         this.settings.link = link;
 
-        let channel = this.master ? 'master' : this.channelNr + 1;
-
-        const msg = {
-            module: 'Soundscape',
-            channel: channel,
-            mode: 'setLink',
-            mute: this.settings.mute,
-            solo: this.settings.solo,
-            link: link
+        if (game.user.isGM) {
+            const payload = {
+              "msgType": "setLink",
+              "channelNr": this.channelNr,
+              link
+            };
+            Hooks.call(moduleName,payload);
         }
-        Hooks.call("Soundscape",msg);
     }
 
     getLink() {
@@ -438,17 +495,21 @@ export class Channel {
         this.audioElement.volume = 1;
 
         this.node = this.context.createMediaElementSource(this.audioElement);
+        
         //this.node.connect(this.context.destination);
         this.configureConnections();
         this.loaded = true;
+        
         if ((wasPlaying && stop == false) || forcePlay) this.play();
 
         let parent = this;
         this.audioElement.addEventListener('loadeddata', function() {
             parent.duration = this.duration;
         });
+  
         this.audioElement.addEventListener('timeupdate', function() {
             if (parent.context.state != 'running') return;
+            
             let timing = parent.settings.timing;
             if (timing == undefined)
                 timing = {
@@ -457,19 +518,60 @@ export class Channel {
                     fadeIn: 0,
                     fadeout: 0
                 }
-            if (parent.playing && parent.audioElement.paused) {
-                if (parent.settings.repeat == 'none') parent.stop();
-                else if (parent.settings.repeat == 'single') {
-                    parent.audioElement.currentTime = timing.startTime;
-                    parent.audioElement.play();
+            let repeat = parent.settings.repeat;
+            if (repeat.minDelay == undefined) {
+                repeat = {
+                    repeat: repeat,
+                    minDelay: 0,
+                    maxDelay: 0
                 }
-                else if (parent.settings.repeat == 'all') parent.next();
+            }
+            let delayTime = repeat.minDelay;
+            if (delayTime == undefined) delayTime = 0;
+            if (repeat.maxDelay != undefined && repeat.maxDelay > repeat.minDelay) {
+                delayTime = Math.random() * (repeat.maxDelay - repeat.minDelay) * 1000;
+            }
+
+            if (parent.playing && parent.audioElement.paused) {
+                if (repeat.repeat == 'none') {
+                    parent.stop();
+                }
+                else if (repeat.repeat == 'single') {
+                    parent.randomizeVolume();
+                   
+                    setTimeout(function(){
+                        parent.audioElement.currentTime = timing.startTime;
+                        if (parent.playing) {
+                            let playPromise = parent.audioElement.play();
+
+                            if (playPromise !== undefined) {
+                                playPromise.then(_ => {
+                                    
+                                })
+                                .catch(error => {
+                                // Auto-play was prevented
+                                });
+                            }
+                        }
+                    },delayTime)
+                }
+                else if (repeat.repeat == 'all') 
+                    parent.randomizeVolume();
+                    setTimeout(function(){
+                        parent.next();
+                    },delayTime)
+                
             }
             if (timing.stopTime > 0 && this.currentTime >= timing.stopTime && parent.fadeStarted == false) {
-                if (parent.settings.repeat == 'none') parent.stop();
-                else if (parent.settings.repeat == 'single') parent.audioElement.currentTime = timing.startTime;
-                else if (parent.settings.repeat == 'all') parent.next();
+                if (repeat.repeat == 'none') parent.stop();
+                else if (repeat.repeat == 'single') {
+                    if (timing.fadeIn > 0 && parent.fadeStarted == false) parent.fade(0,parent.settings.volume,timing.fadeIn)
+                    parent.audioElement.currentTime = timing.startTime;
+                    
+                }
+                else if (repeat.repeat == 'all') parent.next();
             }
+            
             if (timing.fadeOut > 0 && this.currentTime + timing.fadeOut >= timing.stopTime) {
                 if (parent.fadeStarted == false) parent.fade(parent.settings.volume,0,timing.fadeIn)
             }
@@ -477,6 +579,7 @@ export class Channel {
     }
 
     fade(start,end,time) {
+        this.audioElement.volume = start;
         this.fadeStarted = true;
         const stepSize = (end-start)/(time*50);
         let volume = start;
@@ -494,5 +597,17 @@ export class Channel {
                 clearInterval(fade);
             }
         },20)
+    }
+
+    async randomizeVolume() {
+        let volume = this.settings.volume;
+        if (this.settings.randomizeVolume != undefined && this.settings.randomizeVolume > 0) {
+            volume += (Math.random()-0.5)*this.settings.randomizeVolume;
+            if (volume < 0) volume = 0;
+            if (volume > 1.25) volume = 1.25;
+        }
+        
+        if (this.effects.gain != undefined)
+            await this.effects.gain.set(volume);
     }
 }