+++ src/Help/helpMenus.js
@@ -6,7 +6,7 @@ export class helpMenuMixer extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             id: "helpMenu_mixer",
-            title: game.i18n.localize("Soundscape.HelpMenu.Mixer"),
+            title: game.i18n.localize("SOUNDSCAPE.HelpMenu.Mixer"),
             template: "./modules/soundscape/src/Help/helpMenu_mixer.html",
             width: "500px"
         });
@@ -21,7 +21,7 @@ export class helpMenuFx extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             id: "helpMenu_fxConfig",
-            title: game.i18n.localize("Soundscape.HelpMenu.FxConfig"),
+            title: game.i18n.localize("SOUNDSCAPE.HelpMenu.FxConfig"),
             template: "./modules/soundscape/src/Help/helpMenu_fx.html",
             width: "500px"
         });
@@ -36,7 +36,7 @@ export class helpMenuSoundConfig extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             id: "helpMenu_soundConfig",
-            title: game.i18n.localize("Soundscape.HelpMenu.SoundConfig"),
+            title: game.i18n.localize("SOUNDSCAPE.HelpMenu.SoundConfig"),
             template: "./modules/soundscape/src/Help/helpMenu_soundConfig.html",
             width: "500px"
         });
@@ -51,7 +51,7 @@ export class helpMenuSoundboardConfig extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             id: "helpMenu_soundboard",
-            title: game.i18n.localize("Soundscape.HelpMenu.Soundboard"),
+            title: game.i18n.localize("SOUNDSCAPE.HelpMenu.Soundboard"),
             template: "./modules/soundscape/src/Help/helpMenu_soundboardConfig.html",
             width: "500px"
         });
@@ -66,7 +66,7 @@ export class helpMenuSoundscapeConfig extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             id: "helpMenu_soundscape",
-            title: game.i18n.localize("Soundscape.HelpMenu.Soundscape"),
+            title: game.i18n.localize("SOUNDSCAPE.HelpMenu.Soundscape"),
             template: "./modules/soundscape/src/Help/helpMenu_soundscapeConfig.html",
             width: "500px"
         });
@@ -81,7 +81,7 @@ export class helpMenuImport extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             id: "helpMenu_import",
-            title: game.i18n.localize("Soundscape.HelpMenu.Import"),
+            title: game.i18n.localize("SOUNDSCAPE.HelpMenu.Import"),
             template: "./modules/soundscape/src/Help/helpMenu_import.html",
             width: "500px"
         });
@@ -96,7 +96,7 @@ export class helpMenuExport extends FormApplication {
     static get defaultOptions() {
         return mergeObject(super.defaultOptions, {
             id: "helpMenu_export",
-            title: game.i18n.localize("Soundscape.HelpMenu.Export"),
+            title: game.i18n.localize("SOUNDSCAPE.HelpMenu.Export"),
             template: "./modules/soundscape/src/Help/helpMenu_export.html",
             width: "500px"
         });