+++ changelog.md
@@ -1,4 +1,25 @@
 # Changelog
+### v1.0.3 - 01-09-2021
+
+Additions:
+<ul>
+<li>Added support for controlling Soundscape through other modules or macros</li>
+<li>(Potentially) fixed compatibility with Playlist Enhancer</li>
+<li>Added repeat option for the soundboard, this is indicated by a yellow (not playing) or green (playing) shadow around the button</li>
+<li>Repeat for both the soundboard and the mixer now has a configurable delay</li>
+<li>Added option to randomize the volume for the soundboard</li>
+<li>Added 'Apply start timing after first loop' and 'Apply fade in after first loop' to the sound configuration screen. Enabling these will ignore the start timing or fade in when a sound is looped (due to the repeat function) for the first time<li>
+<li>Added option to autostart a soundscape on scene load or when combat starts. Both are configured in the scene configuration under 'Ambience and Atmosphere'.
+</ul>
+
+Fixes:
+<ul>
+<li>Fixed the import function</li>
+<li>Randomizing a playlist or folder did not work</li>
+<li>Fixed CSS issues with Firefox</li>
+<li>Fade in didn't always work, it should be more reliable now</li>
+</ul>
+
 ### v1.0.2 - 14-07-2021
 -Playlists and playlist sounds can not be dragged into the mixer to populare the soundboard or a channel
 -Forgot to fix capitalization issues for the soundboard config, which prevented it from opening